Move source files back to src/ directories, fix broken tests and remove a few outdated references. [t:2h]
