--- src/java/io/riddles/lightriders/game/move/MoveType.java
@@ -1,67 +1,0 @@
-/*
- *  Copyright 2016 riddles.io (developers@riddles.io)
- *
- *      Licensed under the Apache License, Version 2.0 (the "License");
- *      you may not use this file except in compliance with the License.
- *      You may obtain a copy of the License at
- *
- *          http://www.apache.org/licenses/LICENSE-2.0
- *
- *      Unless required by applicable law or agreed to in writing, software
- *      distributed under the License is distributed on an "AS IS" BASIS,
- *      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *      See the License for the specific language governing permissions and
- *      limitations under the License.
- *
- *      For the full copyright and license information, please view the LICENSE
- *      file that was distributed with this source code.
- */
-
-package io.riddles.lightriders.game.move;
-
-import java.awt.*;
-
-/**
- * io.riddles.lightriders.game.move.MoveType
- *
- * [description]
- *
- * @author Joost de Meij - joost@riddles.io, Jim van Eeden - jim@riddles.io
- */
-public enum MoveType {
-    UP,
-    DOWN,
-    LEFT,
-    RIGHT,
-    PASS;
-
-    public MoveType getOpposite() {
-        switch (this) {
-            case UP:
-                return DOWN;
-            case DOWN:
-                return UP;
-            case LEFT:
-                return RIGHT;
-            case RIGHT:
-                return LEFT;
-        }
-
-        return this;
-    }
-
-    public Point getDirection() {
-        switch (this) {
-            case UP:
-                return new Point(0, -1);
-            case DOWN:
-                return new Point(0, 1);
-            case LEFT:
-                return new Point(-1, 0);
-            case RIGHT:
-                return new Point(1, 0);
-        }
-
-        return new Point(0, 0);
-    }
-}
