+++ src/test/resources/bot1_input.txt
@@ -1,0 +1,80 @@
+pass
+down
+garbage
+down
+right
+pass
+right
+up
+right
+right
+up
+up
+up
+up
+right
+right
+down
+right
+down
+down
+up
+up
+up
+up
+right
+right
+down
+right
+down
+down
+up
+up
+up
+up
+right
+right
+down
+right
+down
+down
+up
+up
+up
+up
+right
+right
+down
+right
+down
+down
+up
+up
+up
+up
+right
+right
+down
+right
+down
+down
+up
+up
+up
+up
+right
+right
+down
+right
+down
+down
+up
+up
+up
+up
+right
+right
+down
+right
+down
+down
