+++ src/test/resources/bot2_input.txt
@@ -1,0 +1,80 @@
+left
+left
+left
+down
+down
+down
+down
+left
+left
+left
+left
+left
+left
+left
+left
+left
+left
+up
+up
+up
+up
+pass
+pass
+left
+left
+left
+left
+left
+up
+up
+garbage
+up
+up
+pass
+pass
+left
+left
+left
+left
+left
+up
+up
+up
+up
+pass
+pass
+left
+left
+left
+left
+left
+up
+up
+up
+up
+pass
+pass
+left
+left
+left
+left
+left
+up
+up
+up
+up
+pass
+pass
+left
+left
+left
+left
+left
+up
+up
+up
+up
+pass
+pass
+left
