--- test/resources/bot1_input.txt
@@ -1,80 +1,0 @@
-pass
-move down
-move garbage
-move down
-move right
-pass
-move right
-move up
-move right
-move right
-move up
-move up
-move up
-move up
-move right
-move right
-move down
-move right
-move down
-move down
-move up
-move up
-move up
-move up
-move right
-move right
-move down
-move right
-move down
-move down
-move up
-move up
-move up
-move up
-move right
-move right
-move down
-move right
-move down
-move down
-move up
-move up
-move up
-move move up
-move right
-move right
-move down
-move right
-move down
-move down
-move up
-move up
-move up
-move up
-move right
-move right
-move down
-move right
-move down
-move down
-move up
-move up
-move up
-move up
-move right
-move right
-move down
-move right
-move down
-move down
-move up
-move up
-move up
-move up
-move right
-move right
-move down
-move right
-move down
-move down
