--- test/resources/bot2_input.txt
@@ -1,80 +1,0 @@
-move left
-move left
-move left
-move down
-move down
-move down
-move down
-move left
-move left
-move left
-move left
-move left
-move left
-move left
-move left
-move left
-move left
-move up
-move up
-move up
-move up
-pass
-pass
-move left
-move left
-move left
-move left
-move left
-move up
-move up
-move garbage
-move up
-move up
-pass
-pass
-move left
-move left
-move left
-move left
-move left
-move up
-move up
-move up
-move up
-pass
-pass
-move left
-move left
-move left
-move left
-move left
-move up
-move up
-move up
-move up
-pass
-pass
-move left
-move left
-move left
-move left
-move left
-move up
-move up
-move up
-move up
-pass
-pass
-move left
-move left
-move left
-move left
-move left
-move up
-move up
-move up
-move up
-pass
-pass
-move left
